==> 10_numbers_in_manuscript.Rmd <==
---
title: "Numbers cited in manuscript"
date: "9/30/2021"
output: html_document
---

```{r}
source('_setup.R')

benchmark_version = '2021-05-26'
max_date = "2021-05-19"
```

## Read in data
```{r}
all_polls <- fread(file.path('data', 'final', 'all_polls_all_vars.csv.gz')) %>% 
  filter(pct_error == 0, end_date <= max_date)

all_polls_end_of_march <- all_polls %>% 
  filter(end_date <= '2021-04-01') %>% 
  group_by(mode) %>% 
  filter(end_date == max(end_date))

all_polls_beg_of_may <- all_polls %>% 
  filter(end_date <= '2021-05-10') %>% 
  group_by(mode) %>% 
  filter(end_date == max(end_date))
  
```

# Section 1
## Average sample sizes
```{r}
all_polls %>% 
  filter(pop == 'US', end_date <= max_date) %>% 
  group_by(mode) %>% 
  summarize(avg_sample_size = mean(n), total_n = sum(n))
```

## Differences in willingness estimates by the end of March 2021
```{r}
all_polls_end_of_march %>% 
  filter(mode %in% c('facebook', 'household_pulse'), pop != 'US') %>% 
  pivot_wider(id_cols = c('pop'), names_from = mode, values_from = c(pct_willing)) %>%
  mutate(diff = facebook - household_pulse) %>%
  summarise(min_diff = min(diff), mean_diff = mean(diff), max_diff = max(diff))
```
## Differences in willingness estimates by the end of March 2021
```{r}
all_polls_beg_of_may %>% 
  filter(mode %in% c('facebook', 'household_pulse'), pop != 'US') %>% 
  pivot_wider(id_cols = c('pop'), names_from = mode, values_from = c(pct_willing)) %>%
  mutate(diff = facebook - household_pulse) %>%
  summarise(min_diff = min(diff), mean_diff = mean(diff), max_diff = max(diff))
```

## Rank correlation -- March
```{r}
all_polls_end_of_march_ranked <- all_polls_end_of_march %>%
  filter(mode %in% c('facebook', 'household_pulse'), pop != 'US') %>%
  group_by(mode) %>%
  mutate(rank_willing = frank(-pct_willing), rank_hesitant = frank(-pct_hesitant))

all_polls_end_of_march_ranked %>%
  pivot_wider(id_cols = pop, names_from = mode, values_from = c(pct_hesitant,  pct_willing, pct_vaccinated, pct_pop_vaccinated)) %>%
  summarise(cor_rank_vaccinated = cor(pct_vaccinated_facebook, pct_vaccinated_household_pulse, method = 'kendall')
            , cor_rank_hesitant = cor(pct_hesitant_facebook, pct_hesitant_household_pulse, method = 'kendall')
            , cor_rank_willing = cor(pct_willing_facebook, pct_willing_household_pulse, method = 'kendall')
            , cor_rank_vaccinated_fb_pop = cor(pct_vaccinated_facebook, pct_pop_vaccinated_facebook, method = 'kendall')
            , cor_rank_vaccinated_hp_pop = cor(pct_vaccinated_household_pulse, pct_pop_vaccinated_household_pulse, method = 'kendall')
            )

# all_polls_end_of_march_ranked %>% 
#   filter(pop == 'MO') %>% 
#   mutate(MoE_hesitant = 2 * pct_hesitant_se, lb = pct_hesitant - MoE_hesitant, ub = pct_hesitant + MoE_hesitant) %>% 
#   select(pop, rank_hesitant, rank_hesitant, pct_hesitant , lb, ub)

```

## Rank correlation -- May
```{r}
all_polls_beg_of_may_ranked <- all_polls_beg_of_may %>%
  filter(mode %in% c('facebook', 'household_pulse'), pop != 'US') %>%
  group_by(mode) %>%
  mutate(rank_willing = frank(-pct_willing), rank_hesitant = frank(-pct_hesitant))

all_polls_beg_of_may_ranked %>%
  pivot_wider(id_cols = pop, names_from = mode, values_from = c(pct_hesitant,  pct_willing, pct_vaccinated, pct_pop_vaccinated)) %>%
  summarise(cor_rank_vaccinated = cor(pct_vaccinated_facebook, pct_vaccinated_household_pulse, method = 'kendall')
            , cor_rank_hesitant = cor(pct_hesitant_facebook, pct_hesitant_household_pulse, method = 'kendall')
            , cor_rank_willing = cor(pct_willing_facebook, pct_willing_household_pulse, method = 'kendall'), cor_rank_vaccinated_fb_pop = cor(pct_vaccinated_facebook, pct_pop_vaccinated_facebook, method = 'kendall')
            , cor_rank_vaccinated_hp_pop = cor(pct_vaccinated_household_pulse, pct_pop_vaccinated_household_pulse, method = 'kendall'))

# all_polls_beg_of_may_ranked %>% 
#   filter(pop == 'MO') %>% 
#   mutate(MoE_hesitant = 2 * pct_hesitant_se, lb = pct_hesitant - MoE_hesitant, ub = pct_hesitant + MoE_hesitant) %>% 
#   select(pop, rank_hesitant, rank_hesitant, pct_hesitant , lb, ub)

```



## Increase in Error over time
```{r}
all_polls %>% 
  filter(pop == 'US') %>%
  group_by(mode) %>%
  mutate(first = end_date == min(end_date), last = end_date == max(end_date)) %>%
  filter(first | last) %>%
  select(end_date, n, pct_vaccinated,pct_pop_vaccinated, error)
```

## State-level differences
These are calculated in the plots in `02_plot_fig_1.R`



# Section 2

## Passing 50% uptake
```{r}
all_polls %>%
  filter(pop == 'US', mode %in% c('facebook', 'household_pulse'), end_date %in% c(as.Date('2021-03-27'), as.Date('2021-03-29'))) %>% 
  select(study_name, end_date, pct_vaccinated, ci_2.5, ci_97.5)
  
```


## Design effects
```{r}
all_polls %>%
  filter(pop == 'US') %>%
  group_by(study_name) %>%
  summarize(min(deff), mean(deff), max(deff))
```


# Using the CDC benchmark to identify survey error

```{r}
all_polls %>%
  filter(pop == 'US', end_date > '2021-04-01') %>%
  select(study_name, end_date, ddc_weighted, n_eff_star_cap, pct_reduction_n_eff_cap)
```
